BE/requirements.txt
# ibm-watson-machine-learning==1.0.327
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
requests==2.31.0
# streamlit==1.24.1
# safetensors==0.3.1
# faiss-cpu==1.7.4
flask
ibm-watsonx-ai==1.0.4
elasticsearch==8.12.0